+++ changelog.md
@@ -4,7 +4,6 @@
 
 ### Changed
 
-- Updated settings library and their usage
 - Updated dependencies
 - API tokens will not be exported with the settings
 
@@ -12,6 +11,10 @@
 
 - Fixed an error on the counter during the deletion of a simplified text
 - Fixed a faulty usage of an object type during the process to ignore objects for simplifications
+
+### Removed
+
+- Removed the unused jquery dirty library from the main plugin
 
 ## [3.2.0] - 27.06.2026